Rough Beginning
Defined a class with three lists that will contain the contact information for each person in the list.
Class methods include: print the data contained in the lists, add to the lists.
readData function reads the contacts from an external file and loads to Person class.
new_contact function is a work in progress.